requirements.txt
numpy
torch
torchvision
torchaudio
matplotlib
pillow